ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
tensorflow==2.17.0
scipy==1.14.1
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
